Initial commit and snapshot of how the project looked and worked for the final submission of a semester-long language research project.
There are very minor changes to make sure the project can build in it's current directory structure, but otherwise it's exactly how it was, including having to paste my name, student number, and student email for screenshots alongside code submissions.
From this commit, I will document the process of reviewing, refactoring, and addressing the technical debt.